requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
mlflow
pyyaml
pytest
joblib
httpx